requirements.txt
fastapi
lida
pandas
matplotlib
sqlalchemy
python-dotenv
seaborn
langchain